Fix BoundingBox rotations
Bounding boxes are not transformed correctly with rotation transforms.
This is due to only storing the min and max corners, and assuming that
after rotations the corners still contain the minimum and maximum
coordinates.  This change stores all 4 corners (for 2D; 8 corners for
3D) in the BoundingBox. All of these corner points are transformed, and
we can accurately calculate the min and max as needed (by itemdata).

Fixes #94